requirements.txt
#Python Imports
pyzbar
numpy
web3
qrcode
pillow
